+++ tests/testthat/test-conflicts.R
@@ -4,8 +4,18 @@ test_that("error message is informative", {
   old <- options(crayon.enabled = FALSE)
   on.exit(options(old))
 
-  fun <- conflict_binding("x", c("a", "b", "c"))
-  c <- catch_cnd(fun())
+  c1 <- catch_cnd(conflict_binding("x", c("a", "b", "c"))())
+  c2 <- catch_cnd(conflict_binding("if", c("a", "b", "c"))())
+  c3 <- catch_cnd(conflict_binding("%in%", c("a", "b", "c"))())
 
-  expect_known_output(cat(c$message), test_path("conflicts-error.txt"))
+  expect_known_output(
+    {
+      cat(c1$message)
+      cat("\n\n")
+      cat(c2$message)
+      cat("\n\n")
+      cat(c3$message)
+    }, test_path("conflicts-error.txt"))
 })
+
+