+++ NEWS.md
@@ -4,3 +4,6 @@
 
 * conflicted now listens for `detach()` events and removes conflicts that
   are removed by detaching a package (#5)
+
+* Error messages for infix functions and non-syntactic function names are
+  improved (#14)